# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/dictionaries/{type}": {
      "get": {
        "operationId": "DictionariesController_list",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "departments",
                "managers",
                "priorities",
                "statuses",
                "weights",
                "sizes"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      },
      "post": {
        "operationId": "DictionariesController_create",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "departments",
                "managers",
                "priorities",
                "statuses",
                "weights",
                "sizes"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      }
    },
    "/api/v1/dictionaries/{type}/{id}": {
      "patch": {
        "operationId": "DictionariesController_update",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "departments",
                "managers",
                "priorities",
                "statuses",
                "weights",
                "sizes"
              ],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      },
      "delete": {
        "operationId": "DictionariesController_remove",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "departments",
                "managers",
                "priorities",
                "statuses",
                "weights",
                "sizes"
              ],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      }
    },
    "/api/v1/dictionaries/weights/{id}/apply-to-open-cards": {
      "post": {
        "operationId": "DictionariesController_applyWeight",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      }
    },
    "/api/v1/dictionaries/sizes/recalculate-open-cards": {
      "post": {
        "operationId": "DictionariesController_recalculateSizes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      }
    },
    "/api/v1/custom-fields": {
      "get": {
        "operationId": "CustomFieldsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "custom-fields"
        ]
      },
      "post": {
        "operationId": "CustomFieldsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFieldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "custom-fields"
        ]
      }
    },
    "/api/v1/custom-fields/{id}": {
      "patch": {
        "operationId": "CustomFieldsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFieldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "custom-fields"
        ]
      },
      "delete": {
        "operationId": "CustomFieldsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "custom-fields"
        ]
      }
    },
...
  },
  "info": {
    "title": "PMO Hub API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiSuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string",
            "example": "Операцію виконано"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "success"
        ]
      },
...
      "DictionaryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "capacity_limit_points": {
            "type": "number",
            "minimum": 0
          },
          "department_id": {
            "type": "string",
            "format": "uuid"
          },
          "color": {
            "type": "string",
            "pattern": "^#?([0-9A-F]{3}|[0-9A-F]{4}|[0-9A-F]{6}|[0-9A-F]{8})$"
          },
          "weight": {
            "type": "number",
            "minimum": 0
          },
          "min_score": {
            "type": "number",
            "minimum": 0
          },
          "max_score": {
            "type": "number",
            "minimum": 0
          },
          "is_active": {
            "type": "boolean"
          },
          "code": {
            "type": "string",
            "maxLength": 64
          }
        }
      },
      "CustomFieldDto": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string",
            "enum": [
              "project",
              "task"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "NUMBER",
              "SELECT",
              "CHECKBOX",
              "RICHTEXT"
            ]
          },
          "isRequired": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "showInTable": {
            "type": "boolean"
          },
          "showInCards": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "entityType",
          "name",
          "type",
          "isRequired"
        ]
      }
    }
  }
}